backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
mysqlclient
djangorestframework-simplejwt
django-cors-headers
Pillow
gunicorn
